requirements.txt
groq
langchain_core
langchain-together
langgraph
pyTelegramBotAPI
python-dotenv
